Break out server_manager
